Broken implementation of Backpropagation need fixing But fonctionnal implementation of forward
In this commit we added forward propagation :
as a succession of matrix vector multiplication and vector addition (bias)
and activation function.
Each inputs to the next layer is a vector and for each layer the weights are
stored in a matrix.

And Backward Propagation:
Altough it need fixing due to dimension mismatch , we implemented the core function
of backpropagation :

Backpropagation is implemented  as such :

In order to implement backpropagation we need to find the error at each layer:

This then allow us to find the gradient of the weight and bias of that layer.

We then use those gradient to update the weight .

There is 2 case to calculate the error for each layer:

case 1 We are in the last Layer :
In this case we use a specific formula to compute the error
that will vary based on the type of cost function chosen

case 2 Any other Layer :
In this case we calculate the error based on the previous layer if we are at layer l we calculate
the error using the previously computed l+1 error