Dropout implementation
In this commit we implement dropout in the feed forward mechanism.
We define dropout neuron for each layer and each batch of the training process the modification have mainly been
added to NeuralNetwork::feed_forward.

The dropout rate is now passed down as a parameter in the feed_forward method through the train or train_batch method.